streamlit_requirements.txt
streamlit
pandas
numpy
plotly
python-docx
reportlab
weasyprint
pdfkit
openpyxl